# UbyTecAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "UbyTecAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cliente"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        }
      }
    },
    "/api/Clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Clientes/byusername/{usuario}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        }
      }
    },
    "/api/Clientes/bycedula/{numeroCedula}": {
      "put": {
        "tags": [
          "Clientes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "numeroCedula",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateClienteDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "numeroCedula",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DetallesPedido": {
      "get": {
        "tags": [
          "DetallesPedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DetallePedido"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DetallesPedido"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DetallePedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DetallePedido"
            }
          }
        }
      }
    },
    "/api/DetallesPedido/{id}": {
      "get": {
        "tags": [
          "DetallesPedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DetallePedido"
            }
          }
        }
      },
      "put": {
        "tags": [
          "DetallesPedido"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DetallePedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "DetallesPedido"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DetallesPedido/porCliente/{clienteId}": {
      "get": {
        "tags": [
          "DetallesPedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DetallePedido"
              }
            }
          }
        }
      }
    },
    "/api/DetallesPedido/porAfiliado/{afiliadoId}": {
      "get": {
        "tags": [
          "DetallesPedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "afiliadoId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DetallePedido"
              }
            }
          }
        }
      }
    },
...
    "/api/Pedido/{id}": {
      "get": {
        "tags": [
          "Pedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pedido"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pedido"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Pedido/cliente/{clienteId}": {
      "get": {
        "tags": [
          "Pedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pedido"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Cliente": {
      "type": "object",
      "properties": {
        "clienteID": {
          "format": "int32",
          "type": "integer"
        },
        "numeroCedula": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "apellidos": {
          "type": "string"
        },
        "direccionProvincia": {
          "type": "string"
        },
        "direccionCanton": {
          "type": "string"
        },
        "direccionDistrito": {
          "type": "string"
        },
        "fechaNacimiento": {
          "format": "date-time",
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "usuario": {
          "type": "string"
        },
        "passwordHash": {
          "format": "byte",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DetallePedido": {
      "required": [
        "afiliadoID",
        "clienteID",
        "estado"
      ],
      "type": "object",
      "properties": {
        "detalleID": {
          "format": "int32",
          "type": "integer"
        },
        "afiliadoID": {
          "format": "int32",
          "type": "integer"
        },
        "clienteID": {
          "format": "int32",
          "type": "integer"
        },
        "repartidorID": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "productos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductoDetalle"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Pedido": {
      "type": "object",
      "properties": {
        "pedidoID": {
          "format": "int32",
          "type": "integer"
        },
        "clienteID": {
          "format": "int32",
          "type": "integer"
        },
        "afiliadoID": {
          "format": "int32",
          "type": "integer"
        },
        "repartidorID": {
          "format": "int32",
          "type": "integer"
        },
        "cantidadProducto": {
          "format": "int32",
          "type": "integer"
        },
        "nombreProducto": {
          "type": "string"
        },
        "montoProducto": {
          "format": "double",
          "type": "number"
        },
        "total": {
          "format": "double",
          "type": "number"
        },
        "nombreCliente": {
          "type": "string"
        },
        "nombreComercio": {
          "type": "string"
        },
        "nombreRepartidor": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "ProductoDetalle": {
      "required": [
        "cantidad",
        "detallePedidoID",
        "precio",
        "productoID"
      ],
      "type": "object",
      "properties": {
        "productoDetalleID": {
          "format": "int32",
          "type": "integer"
        },
        "productoID": {
          "format": "int32",
          "type": "integer"
        },
        "detallePedidoID": {
          "format": "int32",
          "type": "integer"
        },
        "cantidad": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 1,
          "type": "integer"
        },
        "precio": {
          "format": "double",
          "minimum": 0,
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateClienteDto": {
      "type": "object",
      "properties": {
        "numeroCedula": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "apellidos": {
          "type": "string"
        },
        "direccionProvincia": {
          "type": "string"
        },
        "direccionCanton": {
          "type": "string"
        },
        "direccionDistrito": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}